Refactor acceptor class to jb::ehs::acceptor.
